requirements.txt
ctransformers
sentence-transformers
pinecone
langchain
langchain-pinecone
langchain-community
langchain-huggingface
flask
pypdf
python-dotenv
-e .